Updated state machine to include further signBitExtension()

+++ X8_multirotor_master/X8_multirotor/ip/AXI_SPI_Driver/AXI_SPI_Driver.h
@@ -110,7 +110,7 @@ struct DTYPE2
  */
 uint16_t xspi_write(uint8_t address, uint8_t val);
 uint16_t xspi_read(uint8_t address);
-int32_t signBitExtend(int16_t testByte);
+int32_t signBitExtend(int32_t testByte);
 
 // main driver
 //void AXI_SPI_DRIVER(volatile int spi_bus[4096], uint32_t pmod_data[4096], uint16_t pmod_test[4096]);
